[myurl] Remove buttons and rename some variables
Summary:
Remove buttons since they are redundant with the presence of links. Links are
more useful in desktop because one can open a link in a tab by middle-click or
control-click, which is handy. In mobile, links and buttons do not matter,
although button is a little bit more intuitive.

Rename variables by removing the "elements" suffix. Use the control name as
the suffix of variables instead.