[myurl] Fix firefox by using indexedDB through indexeddbshim
Summary:
Firefox does not have WebQL, so it breaks. The fix is by using indexed DB
through a polyfill, so the application can be supported by more browsers.

The indexeddbshim is from https://github.com/axemclion/IndexedDBShim.

Fix a bug that if there is an error during deleting of a smart url, the delete
prompt does not close. It's fixed by putting the closing of the prompt (a
dialog) in a finally block.

Test Plan:
1) Remove all smart urls, reload the page, see that all default smart urls are
populated.

2) Add a smart url, reload the page and see the smart url is retained.

3) Delete a smart url, see that it's gone. Reload the page, see that the
smart url is still gone.

4) Add a smart url, delete it, see that it's gone. Reload the page, see that
it's still gone.

Tested with Chrome, Safari and Firefox. The async operations in Safari is
kinda slow, tho.