[set-title] A simple web app to set title of the browser window
Summary:
This simple page is to let you set the title of the page, so that when this
window is minimized in the dock (as in Mac), the title shows up when you hover
the app. This is most useful if you use multiple tabs in a web browser, you
then minimize the window containing the tabs, and you want to what the tabs in
the window are about when you hover the app.

Test Plan:
Go to the page, see the "Current Title" field displays "Set Title"
Set the title to "hello" through the "Set Title", press "Set", see the
"Current Title" field displays "hello". Minimize the window and hover the app
at the dock and see the title is "hello".