[set-title] Manipulate the address bar, so the user can bookmark the page with the title
Summary:
Manipulate the address bar when the user presses the "Set" button, so the
user can bookmark the page with the title.

Add the handling of the query parameter "title" which makes this bookmarking
possible.

Use of window.history.pushState to avoid page reload when the app manipulates
the address bar.

Test Plan:
Set a title "abc", see the address bar changes in additional to the title of
the page. Bookmark this page.

Go to a random website, e.g. google.com. Go to the bookmark from the previous
step, set the title to "aaa 123", see the address bar and the title of the
page changes. Click "back", see they change to "abc". Click "back" again
should go back to the random website, e.g. google.com.