[Note Box][Javascript for Bookmarklets] Placeholder and onchange to onkeyup for Note Box
Summary:
Add a placeholder for the textarea. Use onkeyup event instead of onchange, so
the text is stored right at the time of the user lifting the key instead of
leaving the focus of the textarea. It avoids the situation that the window is
closed unexpectedy while the user is typing, e.g. pressing ctrl+w by
accident...

Also remove the item from localStorage if the text is empty. It's to minimize
the use of localStorage.

Javascript for Bookmarklets should not show 'href="javascript: "' when there
is not input. This commit fixes it.

Test Plan:
For Note Box, see the placeholder appears. Enter some text, presses ctrl+w,
reopen the window/tab should see the text is restored. Also use the Chrome
console to check that while there is no text in the
sorno-extra-box-textarea-input, there should not be an item
"sorno-extra-box-textarea-stored-text" in localStorage.

In Javascript for Bookmarklets, enter some text, then delete the text to see
that no 'href=...' appears in output.