[timer] Show the time that the timer went off
Summary:
Show the time that the timer went off when the timer goes off.

Fix a bug that re-setting a timer does not show the time lapsed. It's caused
by a previous timer went off that might toggle the output element to off and
set the opacity to 0.

Rename the variables to make the code less confusing about the "time"
variables are for Date objects or for milliseconds.